requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
easyocr==1.7.2
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.1
ollama==0.6.1
opencv-python-headless==4.13.0.90
packaging==25.0
pillow==12.1.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-bidi==0.6.7
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.0
shapely==2.1.2
sympy==1.14.0
tifffile==2026.1.14
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
